Created New Screen called Budget Planner
creates a new screen that contains all the main parts of the program. Asks user to input values such as store name, item, and price for their organization purposes. User will also be asked to enter a budget that they want to maintain. Program will add their daily spending and notify them if they are on track according to their goal.